refactor : null 포인터 exception 해결을 위해, getTwoScores 메소드를 추상화 하여 nomal과 Final로 동작을 나눔
